updated movement facade and added outside block manager
