+++ PythonScripts/Movement/MovementFacade.py
@@ -1,14 +1,17 @@
 from MoveImpl import MoveImpl
 from InterruptImpl import InterruptImpl
+from ..OutsideBlockFacade import OutsideBlockFacade
+from Enums.Board import Board
 
 
 class MoveFacade:
     def __init__(self):
+        self.block_manager = OutsideBlockFacade()
         self.move = MoveImpl()
         self.interrupt = InterruptImpl()
 
-    def do_move_block(self, frm, to):
-        return self.move.do_move_block(frm, to)
+    def do_move_block(self, to: Board):
+        return self.move.move_object(pos_from=self.block_manager.take_block(), pos_to=to.value)
 
     def do_interrupt_move(self):
         return self.interrupt.do_interrupt_move()
